support new checker schema
